+++ tests/manual/proteindatabank/proteindatabanktest.cpp
@@ -52,7 +52,7 @@ void ProteinDataBankTest::downloadFile()
     QCOMPARE(file->polymerCount(), size_t(1));
     chemkit::Polymer *protein = file->polymer();
     QCOMPARE(protein->chainCount(), size_t(1));
-    QCOMPARE(protein->chain(0)->residueCount(), 129);
+    QCOMPARE(protein->chain(0)->residueCount(), size_t(129));
 
     delete file;
 }
